Optimize the If node to facilitate case nodes better
Since we don't have a `case` node anymore, we can make an `if` node
generate code better.

This commit uses a special property that `$truthy($truthy(a) ? true : b)`
is the same as `$truthy($truthy(a) || b)` which itself is the same as
`$truthy(a) || $truthy(b)`. Similarly we do with false on the falsy side
now, but true on truthy is what `case a, b` generates actually. And
`$truthy((a)['$==='](b))` we convert to `$eqeqeq('a', 'b')`. So - this
commit makes a lot more `$eqeqeq`s. And `$eqeqeq` may be optimized akin
to how `$rb_plus` is optimized.

But the most important thing is that it simplifies the JS output, which
will make reading things a lot easier.

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 5.962 -> 5.947 (change: -0.24%)
                      Run time: 0.269 -> 0.269 (change: -0.31%)
                   Bundle size: 4784558 -> 4765596 (change: -0.40%)
          Minified bundle size: 1027822 -> 1015690 (change: -1.18%)
            Mangled & minified: 716587 -> 702179 (change: -2.01%)
```