Don't generate $ret_or_ tmp values conditionally.
Sometimes we don't care about the return value except of if it's
truthy or falsy.

Let's consider this example, real-life code:

```ruby
    elsif self < 0 && (::Float === other || ::Rational === other)
```

It used to generate the following:

```javascript
      } else if ($truthy(($truthy(($ret_or_1 = $rb_lt(self, 0))) ? (($truthy(($ret_or_2 = $$$('Float')['$==='](other))) ? ($ret_or_2) : ($$$('Rational')['$==='](other)))) : ($ret_or_1)))) {
```

If we don't create the ret_or variables, we may use the optimizations
we applied before. After this patch, we generate the following:

```javascript
  } else if (($rb_lt(self, 0) && (($eqeqeq($$$('Float'), self.$other()) || ($eqeqeq($$$('Rational'), self.$other())))))) {
```

The optimization, based on the current master, gives us the following:

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 5.968 -> 5.927 (change: -0.69%)
                      Run time: 0.270 -> 0.264 (change: -1.98%)
                   Bundle size: 4784558 -> 4697937 (change: -1.81%)
          Minified bundle size: 1027822 -> 999525 (change: -2.75%)
            Mangled & minified: 716587 -> 697985 (change: -2.60%)
```