Optimize $truthy and $not
This came from my cooperation with Elia on writing extensive benchmarks
for how to write the $truthy helper in the best way for performance
optimization.

While I heavily support removing support for boxed booleans, we are not
doing it yet. Support for those incurs some overhead still, this commit
brings us a lot closer to the desired performance.

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.089 -> 5.942 (change: -2.41%)
                      Run time: 0.269 -> 0.264 (change: -1.84%)
                   Bundle size: 4784558 -> 4665609 (change: -2.49%)
          Minified bundle size: 1027822 -> 998545 (change: -2.85%)
            Mangled & minified: 716587 -> 697179 (change: -2.71%)
```

Co-authored-by: Elia Schito <elia@schito.me>