Try to compile case statements as switch
This improves deeply nested `case` structures, like in... lexer.
Compared to a tree of else ifs, `switch` is more like a goto, so more
likely to be more performant.

This makes our `ifs` (or `&&`, `||`) to be compiled to either `if`,
ternary, `&&`, `||` or `switch`, making use of all the JS control flow
features.