Proc#call: Refactor for performance
Current Proc#call implementation has several conditional branches and
a try/catch block, all of which might impact performance. The main
logic is to call a block, and the exceptional scenarios are when
$$brk or $$ret is thrown. The current implementation always checks
for these exceptions which could be avoided in the common case.

An optimized version that reduces the amount of branching and avoids
setting up a try/catch block unless needed.

In this version, we've factored out the invocation of a proc or a
lambda into helper functions $call_proc and $call_lambda. In each
function, we set up the try/catch block only if necessary (i.e.,
if $$brk or $$ret is present). In Proc#call, we simply check if
a block is passed and attach it to the proc. Then we delegate to
the appropriate function based on whether the proc is a lambda.

The main goal of this refactor is to streamline the common path
and remove unnecessary branching and exception handling overhead.

(ChatGPT4)

$ bundle exec rake bench:ips FILE=benchmark-ips/bm_block_vs_yield.rb
before:              block:  3033756.5 i/s - 1.30x  slower
after:               block:  3272548.5 i/s - 1.26x  slower

The difference is small, but tangible and resulting code looks much
cleaner.