Hash: Optimize #to_a
Pre-allocate the memory for the result array, which can help performance
when dealing with very large hashes.

Pre-allocating memory is beneficial because it prevents the JavaScript
engine from having to continuously resize the array as elements are added.

In this version, we create an array with a predefined size (equal to the
number of keys in the hash) using new Array(length). Then, instead of
pushing elements into the array, we assign each key-value pair directly
to the corresponding index in the result array.

(ChatGPT4)