Bridging: call original constructor for subclasses of level larger than 1
This commit introduces a function `descends_from_bridged_class(klass)` to check
if a class or any of its ancestors is a bridged class. During the allocation of
a new class, the `descends_from_bridged_class()` function is now used to
determine if the new class is a descendant of a bridged class, and if so, it
ensures the original JavaScript constructor is correctly called. This enables
more accurate and robust handling of class hierarchies involving bridged classes.

This problem has manifested in Firefox which didn't capture stacktraces for
RuntimeError, but did so for StandardError.

This fixes #2540