iclass: Propagate removal of method from including/prepending modules.
How does including/prepending a module work in Opal in short -
we create a new JavaScript "class" so to speak, called an `iclass`.
We copy the methods from a module to that `iclass` and then we
inject that `iclass` into the prototype chain.

Unfortunately, before this commit, if a method was removed from
a module, an iclass wouldn't get updated.

This was needed for Flexmock to work under Opal. I believe this
will also allow us to make RSpec-Mocks to pass a lot more tests
in Opal-RSpec.

I am particularly excited about this fix :D

This PR has been sponsored by Ribose Inc. ref: plurimath/plurimath#159