Fix Typo in `Running tests` Section of `README.md` File
Removed the duplicated word `will` from the last sentence of `Running tests` section, which's inside `README.md` file.

Made the `Automated runs` Sub-section of `Running test` have the same
Heading Level as `MSpec` and `RSpec` Sub-sections, which's Level 3
Heading (###).
And moved `Automated runs` to be the last Sub-section of `Running tests`
Section, to make the Flow of Reading more "digestible" to New-Readers of
opal.

See the commit patch(es) for exact info on changes.