Opal::Project
This commit creates a concept of a project, which is a concept similar
to a Gem, but rather for JavaScript parts.

A single Gem can contain multiple projects. A project is most often
defined by adding an Opalfile file that contains an Opalfile DSL.
A project is added to compilation by adding an appropriate require
path or by compiling a file,

As of now, this DSL contains just two directives:
- add_gem_dependency - declares that files from a given gem should be
                       available for use by Opal's require directive
- add_load_path - adds a load path. This is useful for Opal gems,
                  so that you could specify that eg. "opal" directory
                  contains Opal files and "lib" contains Ruby files.